+++ Components/Audio/Events/AudioCommandSetOutputDevice.cs
@@ -7,7 +7,7 @@ namespace Slipstream.Components.Audio.Events
     {
         public string EventType => "AudioCommandSetOutputDevice";
         public bool ExcludeFromTxrx => false;
-        public string PluginId { get; set; } = "INVALID-PLUGIN-ID";
+        public string InstanceId { get; set; } = "INVALID-INSTANCE-ID";
         public int DeviceIdx { get; set; } = -1;
         public ulong Uptime { get; set; }
 
@@ -16,7 +16,7 @@ namespace Slipstream.Components.Audio.Events
             return obj is AudioCommandSetOutputDevice device &&
                    EventType == device.EventType &&
                    ExcludeFromTxrx == device.ExcludeFromTxrx &&
-                   PluginId == device.PluginId &&
+                   InstanceId == device.InstanceId &&
                    DeviceIdx == device.DeviceIdx;
         }
 
@@ -25,7 +25,7 @@ namespace Slipstream.Components.Audio.Events
             int hashCode = 1390535527;
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(EventType);
             hashCode = hashCode * -1521134295 + ExcludeFromTxrx.GetHashCode();
-            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(PluginId);
+            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(InstanceId);
             hashCode = hashCode * -1521134295 + DeviceIdx.GetHashCode();
             return hashCode;
         }