+++ Components/Audio/Events/AudioOutputDevice.cs
@@ -9,7 +9,7 @@ namespace Slipstream.Components.Audio.Events
     {
         public string EventType => "AudioOutputDevice";
         public bool ExcludeFromTxrx => false;
-        public string PluginId { get; set; } = "INVAILD-PLUGIN-ID";
+        public string InstanceId { get; set; } = "INVAILD-INSTANCE-ID";
         public string Product { get; set; } = string.Empty;
         public int DeviceIdx { get; set; } = -1;
         public ulong Uptime { get; set; }
@@ -19,7 +19,7 @@ namespace Slipstream.Components.Audio.Events
             return obj is AudioOutputDevice device &&
                    EventType == device.EventType &&
                    ExcludeFromTxrx == device.ExcludeFromTxrx &&
-                   PluginId == device.PluginId &&
+                   InstanceId == device.InstanceId &&
                    Product == device.Product &&
                    DeviceIdx == device.DeviceIdx;
         }
@@ -29,7 +29,7 @@ namespace Slipstream.Components.Audio.Events
             int hashCode = -1602761893;
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(EventType);
             hashCode = hashCode * -1521134295 + ExcludeFromTxrx.GetHashCode();
-            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(PluginId);
+            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(InstanceId);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(Product);
             hashCode = hashCode * -1521134295 + DeviceIdx.GetHashCode();
             return hashCode;