+++ Components/Audio/Lua/IAudioInstanceThread.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+
+using System;
+
+namespace Slipstream.Components.Audio.Lua
+{
+    public interface IAudioInstanceThread : IDisposable
+    {
+        void Start();
+    }
+}