+++ Components/BasePlugin.cs
@@ -31,7 +31,7 @@ namespace Slipstream.Components
 
         public IEventHandlerController EventHandlerController { get; }
 
-        public BasePlugin(IEventHandlerController eventHandlerController, string id, string name, string displayName, bool reconfigurable = false, bool fullThreadControl = false)
+        protected BasePlugin(IEventHandlerController eventHandlerController, string id, string name, string displayName, bool reconfigurable = false, bool fullThreadControl = false)
         {
             EventHandlerController = eventHandlerController;
             Id = id;