+++ Components/Internal/ILuaService.cs
@@ -1,8 +1,10 @@
 ﻿#nullable enable
 
+using System;
+
 namespace Slipstream.Components.Internal
 {
-    public interface ILuaService
+    public interface ILuaService : IDisposable
     {
         ILuaContext Parse(string filename);