+++ LuaScripts/iracing_prechecks.lua
@@ -1,12 +1,14 @@
 -- At start of qualify and race, remind user to check their car setup
 
+local audio = require("api/audio"):instance({ id = "audio" })
+
 iracing:send_session_state()
 
 function handle(event)
 	if event.EventType == "IRacingQualify" and event.State == "Racing" then
-		audio:say("AudioPlugin", "Qualify checklist: Setup and fuel", 1.0)
+		audio:say("Qualify checklist: Setup and fuel")
 	end
 	if event.EventType == "IRacingRace" and event.State == "GetInCar" then
-		audio:say("AudioPlugin", "Race checklist: Setup, tyres, fuel and fast-repairs", 1.0)
+		audio:say("Race checklist: Setup, tyres, fuel and fast-repairs")
 	end
 end