+++ Shared/ILuaLibraryRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace Slipstream.Shared
+{
+    public interface ILuaLibraryRepository
+    {
+        public ILuaLibrary Get(string name);
+    }
+}