+++ Shared/ILuaReference.cs
@@ -1,0 +1,10 @@
+﻿#nullable enable
+
+using System;
+
+namespace Slipstream.Shared
+{
+    public interface ILuaReference : IDisposable
+    {
+    }
+}